More saddlebag work - Establish null device for invalid saddlebag indices - Minor changes to sb_setAllAmps for compatibility with other "all"   commands

+++ argus.h
@@ -87,7 +87,7 @@ extern int  sb_ampPow(char *inp, int sbNum);
 extern int  sb_ledOnOff(char *inp, int sbNum);
 extern int  sb_readADC(int sbNum);
 extern BYTE sb_readPLLmon(int sbNum);
-extern int  sb_setAllAmps(int v);
+extern int  sb_setAllAmps(float v);
 
 /**************************************************************************/
